# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample of RESTful api made by cargo-rest-api",
        "title": "cargo-rest-api",
        "contact": {
            "name": "Aleksand Efremov",
            "url": "https://github.com/efremovich/cargo-rest-api",
            "email": "sasha.fima@gmail.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8888",
    "paths": {
...
        "/api/v1/external/sities": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Get list of existing sities.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sities"
                ],
                "summary": "Get sities",
                "parameters": [
                    {
                        "enum": [
                            "en",
                            "ru"
                        ],
                        "type": "string",
                        "default": "en",
                        "description": "Language code",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Request id",
                        "name": "Set-Request-Id",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.successOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Create a new sity.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sities"
                ],
                "summary": "Create a new sity",
                "parameters": [
                    {
                        "enum": [
                            "en",
                            "ru"
                        ],
                        "type": "string",
                        "default": "en",
                        "description": "Language code",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Request id",
                        "name": "Set-Request-Id",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Sity name",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Sity latitude",
                        "name": "latitude",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Sity longitude",
                        "name": "longitude",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.successOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    }
                }
            }
        },
        "/api/v1/external/sities/uuid": {
            "put": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Update an existing sity.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sities"
                ],
                "summary": "Update sity",
                "parameters": [
                    {
                        "enum": [
                            "en",
                            "ru"
                        ],
                        "type": "string",
                        "default": "en",
                        "description": "Language code",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Request id",
                        "name": "Set-Request-Id",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Sity UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Sity name",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Sity latitude",
                        "name": "latitude",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Sity longitude",
                        "name": "longitude",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.successOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    }
                }
            }
        },
        "/api/v1/external/sities/{uuid}": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Get detail of existing sity.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sities"
                ],
                "summary": "Get sity",
                "parameters": [
                    {
                        "enum": [
                            "en",
                            "ru"
                        ],
                        "type": "string",
                        "default": "en",
                        "description": "Language code",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Request id",
                        "name": "Set-Request-Id",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Sity UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.successOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Delete an existing sity.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sities"
                ],
                "summary": "Delete sity",
                "parameters": [
                    {
                        "enum": [
                            "en",
                            "ru"
                        ],
                        "type": "string",
                        "default": "en",
                        "description": "Language code",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Request id",
                        "name": "Set-Request-Id",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Sity UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.successOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.errorOutput": {
            "type": "object",
            "properties": {
                "args": {
                    "type": "string"
                },
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error_code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.successOutput": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "meta": {}
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "api_key",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "JWTAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "Oauth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "http://localhost:8888/oauth/authorize",
            "tokenUrl": "http://localhost:8888/oauth/token",
            "scopes": {
                "all": " Grants all access"
            }
        },
        "Oauth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "http://localhost:8888/oauth/token",
            "scopes": {
                "all": " Grants all access"
            }
        }
    }
}